Rename to tiptap-iframe
